style(ui): refresh app layout
