+++ aitutor/pages/manage_exercises/page.py
@@ -19,6 +19,37 @@ from aitutor.pages.navbar_admin import with_admin_navbar
 from aitutor.utilities.filtering_components import search_badges, search_bar
 
 
+def exercise_toolbar() -> rx.Component:
+    """Render grouped actions for managing exercises."""
+    return rx.box(
+        rx.vstack(
+            rx.hstack(
+                rx.hstack(
+                    import_exercises_button(),
+                    add_exercise_button(),
+                    edit_tags_button(),
+                    class_name="exercise-toolbar-group",
+                ),
+                rx.hstack(
+                    delete_selected_exercises_button(),
+                    export_selected_exercises_button(),
+                    class_name="exercise-toolbar-group",
+                ),
+                class_name="exercise-toolbar-row",
+                width="100%",
+            ),
+            rx.hstack(
+                search_bar(ManageExercisesState),
+                class_name="exercise-search-row",
+                width="100%",
+            ),
+            spacing="3",
+            width="100%",
+        ),
+        class_name="exercise-toolbar",
+    )
+
+
 @with_navbar(routes.ADMIN_SETTINGS)
 @with_admin_navbar(routes.MANAGE_EXERCISES)
 @page_require_role_or_permission(required_role=UserRole.ADMIN)
@@ -26,58 +57,13 @@ def manage_exercises_page() -> rx.Component:
     """Manage exercises page."""
     return rx.center(
         rx.vstack(
-            rx.desktop_only(
-                rx.hstack(
-                    rx.hstack(
-                        delete_selected_exercises_button(),
-                        export_selected_exercises_button(),
-                    ),
-                    rx.hstack(
-                        import_exercises_button(),
-                        add_exercise_button(),
-                    ),
-                    align="center",
-                    justify="between",
-                    width="100%",
-                ),
-                width="100%",
-            ),
-            rx.mobile_and_tablet(
-                rx.hstack(
-                    import_exercises_button(),
-                    add_exercise_button(),
-                ),
-            ),
-            rx.mobile_and_tablet(
-                rx.hstack(
-                    delete_selected_exercises_button(),
-                    export_selected_exercises_button(),
-                ),
-            ),
-            rx.mobile_and_tablet(
-                edit_tags_button(),
-            ),
-            rx.mobile_and_tablet(
-                search_bar(ManageExercisesState),
-            ),
-            rx.desktop_only(
-                rx.hstack(
-                    search_bar(ManageExercisesState),
-                    edit_tags_button(),
-                    align="center",
-                    justify="between",
-                    width="100%",
-                ),
-                width="100%",
-            ),
+            exercise_toolbar(),
             search_badges(ManageExercisesState),
             exercise_table(),
             spacing="3",
             align="center",
-            justify="center",
+            justify="start",
             width="100%",
         ),
-        margin_top="2em",
-        margin_bottom="2em",
-        width="100%",
+        class_name="page-frame",
     )
