+++ aitutor/pages/manage_users/page.py
@@ -26,7 +26,5 @@ def manage_users_page() -> rx.Component:
             align="center",
             justify="center",
         ),
-        margin_top="2em",
-        margin_bottom="2em",
-        width="100%",
+        class_name="page-frame",
     )
