+++ aitutor/pages/reports/page.py
@@ -25,7 +25,5 @@ def reports_page() -> rx.Component:
             align="center",
             justify="center",
         ),
-        margin_top="2em",
-        margin_bottom="2em",
-        width="100%",
+        class_name="page-frame",
     )
